refactor DirItem to use PathBuf
